converting to dynamic

+++ models/user.js
@@ -48,6 +48,18 @@ module.exports = (sequelize, Sequelize) => {
                 type: Sequelize.STRING,
                 allowNull: false,
             },
+            allow: {
+                type: Sequelize.ENUM("Allow", "Block"),
+                allowNull: false,
+            },
+            totalProfit: {
+                type: Sequelize.STRING,
+                alllowNull: false,
+            },
+            commission: {
+                type: Sequelize.DOUBLE(20, 6),
+                allowNull: false
+            }
         },
         {
             tableName: "users",
@@ -59,7 +71,7 @@ module.exports = (sequelize, Sequelize) => {
     User.migrate = async () => {
         const count = await User.count();
 
-        // if (!count) {
+        if (!count) {
             await User.destroy({ truncate: true });
             const saltRounds = 10;
             const hashedPassword = await bcrypt.hash("123456", saltRounds);
@@ -71,7 +83,7 @@ module.exports = (sequelize, Sequelize) => {
                 server: 'trading',
                 leverage: 1                                  /////////////////////////
             })
-        // }
+        }
     };
 
     return User;
